+++ colander/tests/test_colander.py
@@ -171,6 +171,21 @@ class TestInvalid(unittest.TestCase):
                 },
             )
 
+    def test_asdict_with_msg_none(self):
+        # see https://github.com/Pylons/colander/pull/333
+        from colander import All
+
+        node1 = DummySchemaNode(None, 'node1')
+        validator1 = DummyValidator('validator1')
+        validator2 = DummyValidatorWithMsgNone()
+        validator = All(validator1, validator2)
+        exc = invalid_exc(validator, node1, None)
+        try:
+            d = exc.asdict()
+        except TypeError as error:
+            # TypeError: sequence item 1: expected str instance, NoneType found
+            self.fail(str(error))
+
     def test___str__(self):
         from colander import Positional
 
@@ -4671,6 +4686,14 @@ class DummyValidator(object):
             e = Invalid(node, self.msg)
             self.children and e.children.extend(self.children)
             raise e
+
+
+class DummyValidatorWithMsgNone(object):
+    def __call__(self, node, value):
+        from colander import Invalid
+
+        e = Invalid(node)
+        raise e
 
 
 class Uncooperative(object):